Allow JitDump compilations to always get queued
The logic in `compileOnSeparateThread` attempts to prevent compilations
to happen if there are no available compile threads or if further
compilations have been disabled by outside factors. This is a reasonable
request, however a request which should not hold true for JitDump
compilations. For JitDump compilation the JitDump logic will have
disabled further compilations, purged the compile queue, and have
activated the diagnostic compilation thread.

In the time between activating these events (which are performed on a
crashed thread) the active compilation threads could have attempted
to process their queues and self-suspended because the queues have
been purged by the JitDump logic. Thus we could reach the location being
changed in this commit with all "normal" compilation threads being
suspended, meaning `getNumCompThreadsActive() == 0` and further
compilations have been disabled via the JitDump process, however we
still want to queue the compilation because it is a JitDump request and
the diagnostic thread will be processing this compilation entry.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>