Avoid starting suspended threads for JitDump compilations
When attempting to add a new method to the compilation queue during the
JitDump process we can encounter a situation in which the logic inside
of `shouldActivateNewCompThread` tells us that we should activate a new
compilation thread to handle the compilations.

This is because during the JitDump process we have marked the JIT to
disable further compilations, and purged the compilation queue. When
this is done, compilation threads will self-suspend themselves as they
attempt to process new queue entries and encounter that the queue is
actually empty.

In a worse case the interpreter could have signaled VM shutdown in
which case we would call `stopCompilationThreads` which would singal
all threads to terminate, but the crashed thread currently processing
the JitDump logic will not be terminated of course since it cannot be
interrupted. In such a case the logic in `stopCompilationThreads`
looks at the number of threads currently active and it sees there are
in fact 0, and it returns `TR_yes` to activate a new thread. Subsequent
code then calls `getFirstSuspendedCompilationThread` thinking there must
be a thread available, but there isn't because they are all (but one)
terminated! It will return a NULL result, and we will crash trying to
call `resumeCompilationThread`. The JitDump singal handler will catch
this crash thinking it was a recursive crash and we will not generate
the recompilation trace log (jitdump).

We fix this by returning `TR_no` if further compilations are disabled,
something that is always set by the JitDump process.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>