Reacquire VM access if we have a recursive crash during JitDump
We have acquired VM access above and will be releasing it below. It is
possible however that during the compilation process in the above
`compile(...)` call we have released VM access and have subsequently
crashed or asserted. In such a scenarion the JitDump process will have
started and queued the method for recompilation with a custom signal
handler to catch the crash/assertion and return. The JitDump
compilation will have hopefully crashed/asserted in the same place as
the original compilation, and thus we will not be holding VM access at
that point. This means we will return from the above call to
`compile(...)` without holding VM access, and below we will attempt to
call `j9jni_deleteLocalRef` tounpin the class. The
`j9jni_deleteLocalRef` function has a requirement that we hold VM
access otherwise it will assert. We do not want to assert, as that will
end up calling `abort()` and the JitDump process will not complete, nor
will any dump handlers following JitDump.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>