Enable JitDump tracing options based off of method request
Previously, the decision on whether to enable tracing options for a
JitDump method compilation request was based off of whether it is the
diagnostic thread which ended up processing the `JitDumpMethodDetails`
entry. While this should always be the case, and we will fix this in
subsequent commits, the decision on whether to enable tracing should
solely be based on whether it is a JitDump compile, no which thread is
processing it.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>