Ensure compilation threads have stopped before stopping diagnostic thread
In the event that the interpreter shudown has triggered, causing us to
request that all compilation threads be stopped we want to ensure that
the non-diagnostic threads are stopped first. This avoids the situation
in which a compilation thread has crashed and is processing the JitDump
request while the `stopCompilationThreads` function is called by the VM.

In such an event the diagnostic thread can be terminated before being
resumed by the JitDump process, and we could block forever waiting for
the JitDump compilation request to be complete, because there is no
active thread ready to process it.

To solve this problem we ensure all non-diagnostic threads have
terminated fully before terminating the diagnostic threads. This ensures
that even if a compilation thread did crash and is processing the JitDump
we will not terminate the diagnostic threads, because the crashed thread
is still considered active.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>