Fix incorrect access of Compilation object
In the catch block of `TR::CompilationInfoPerThreadBase::compile`
we were accessing the compilation object and calling its destructor.
However, by that point the heap memory region is out of scope and thus
the pointer to the compilation object is invalid.
This bug can be triggered on JITServer, if adding hints at the
end of the compilation results in `StreamFailure` exception.
To handle it, this commit surrounds code that adds hints with
try/catch block and fails the compilation if stream failure happens.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>