Diagnostic thread should process an entry if and only if it is a JitDump compile
Only the diagnostic thread should be processing JitDump compilations.
This is to ensure proper tracing is active and proper synchronization
is performed w.r.t. special events (interpreter shutdown, etc.). The
logic here is split into two parts. In the event we are a diagnostic
thread, the JitDump process will have ensured the method compilation
queue has been purged and all further compilations have been suspended
until the entire JitDump process in complete. This also means if we are
a diagnostic thread, then the only entries in the queue should be
JitDump compile requests. We will ensure this is the case via a fatal
assert.

In addition there can be scenarios in which a non-diagnostic
compilation thread is still attempting to process entries while the
JitDump process (in a crashed thread) is happening. This is because one
compilation thread must always be active, and there is a timing hole
between purging of the queue in the JitDump process, and when the
diagnostic thread is resumed. This means a non-diagnostic thread could
attempt to pick up a JitDump compilation request (see
eclipse/openj9#11772 for details). We must ensure that this does not
happen and simply skip the request if we are a non-diagnostic thread
attempting to process a JitDump compilation.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>