Catch SIGTRAP signals from compilations
We changed the signal that is raised via `TR::trap()` in OMR to raise
`SIGTRAP` instead of `SIGABRT`. We want to catch such signals to
generate the diagnostic output and dumps as well as to catch the
recursive traps which may be issued on the diagnostic threads and we
want to handle such traps properly by ignoring them.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>