Check DF flag in X86 signal handler
In the event of a seg fault in the JIT code, the DF flag should be
cleared, otherwise something must be wrong in the JIT code, in which
case we should fail the VM right away to generate a core dump with the
right context for defect investigation.

This commit adds such check which is to be enabled by an environment
variable in testing.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>