Fix a faulty assert related to number of active threads
It is possible that we have disabled further compilations due to an
event such as running out of code cache. In such events the "disable
further compilation" flag will be set and no new compilation threads
should be activated, because there won't be anything to process.

We had an assert in `resumeCompilationThread()` which asserts that at
the end of the function we have at least one active thread. However
as just explained this may not be the case always.

The assert is removed and replaced with a verbose log message so that
we can still look out for this scenario.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>