Update mechanism for checking invokedynamic/handle is resolved
When unresolved, the adapter method symbol constructed is
linkToStatic, which requires 2 additional args on top of
the original ROM Method parameters for invokedynamic and
3 additional args in the case of invokehandle. In both
cases, when resolved, there would be one less invoke cache
array element load required as opposed to the unresolved
case. This way of checking whether we are dealing with
an unresolved case would ensure that the right number of
objects are pushed even if resolution happens in between
creating the target method and loading the invoke cache
elements.

The new mechanism relies on passing a pointer to a boolean
variable from ILGen that gets set before accessing the call
site table or constant pool table.

Furthermore, comments have been added in ILGen to
demonstrate the different kinds of call nodes created for
resolved and unresolved invokedynamic and invokehandle
bytecodes.

Included as part of this changeset are some additional
clarifications in existing comments as well as comments/checks
where needed for parts of the new bytecode handler implementation
that would not work in AOT and/or JITServer.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>