Handle potential exception case during copyforward completeScan
Copyforward/markCompact hybrid collector is the default PGC collector.
it can retrieve scan work from both scanCache and Work Packet, But
for the best performance, we only try to retrieve the works from
Work Packet for the abort case or the live objects in non evacuation
regions. for extremely case, the scan work might be retrieved from
work packet during copyforward to markcompact transition, so we should
handle this case as well.

Signed-off-by: Lin Hu <linhu@ca.ibm.com>