Don't AOT compile a method already in the SCC
When using loadLimit/LoadLimitFile, a method body that's already in the
SCC will not be loaded. However, because the check for AOT compilation
eligibility does not bother checking whether the method body is already
in the SCC, it will result in an unnecessary AOT warm. This is not only
inefficient (as the method body would not be loaded) but bad for
performance because when storeCompiledMethod is called, because a method
body already exists in the SCC, it will mark that body as stale. This
means that that method body will never be loaded again.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>