Fix a SEGV executing HashMap$HashIterator.nextNode() concurrently.
When using a HashMap concurrently it's possible for an SEGV to occur
if two threads are executing HashMap$HashIterator.nextNode(). The
best solution is to insure that HashMap is only used by a single
thread or is externally synchronized. This fix will prevent the JIT
from removing bound checks for HashMap$HashIterator.nextNode().
With this fix the SEGV will be replaced by an
ArrayIndexOutOfBoundsException

Signed-off-by: Kevin Langman <langman@ca.ibm.com>