Enable Global Lock Reservation by default on Power
Global Lock Reservation is now enabled by default on Power. This is the same
behavior as if the -XX:+GlobalLockReservation option was used. To use the
previous form of lock reservation, Global Lock Reservation can be disabled
with the option -XX:-GlobalLockReservation.

Global Lock Reservation is not supported on other platforms so this change
only affects Power.

Updated documentation and comments to reflect Global Lock Reservation now
being on by default on Power.

Issue: #11424
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>