Decompress String before invoking decompressedArrayCopy
In the private String(String,Char) constructor, it is possible
that to end up in the else block where the String is
compressed, but the char > 255. In such a case, the String needs
to be decompressed correctly before being copied over. This
commit ensures that is done correctly now. A test is also added
to catch this corner case now.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>