Move VT lowering code to new TreeLowering optimization
This commit moves code related to Valhalla value types from lowerTrees
to the new tree lowering optimization. Doing so will eliminate around
bad interactions between the different parts of lowerTrees and create
the opportunity to clean up the trees a little post lowering.

Modifications to the code are only the minimum necessary to adapt it to
the structure of the new optimization.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>