Try to ensure the same options are used for jitdump compilations
We have the original `TR::Options` from the crashed compilation. The
above logic which adjusts various options is timing sensitive, and
can be dependent on the VM state (startup vs not), which can
drastically change how the compilation looks (ex. is NextGenHCR
enabled, is SVM to be used).

Ideally we would just copy construct the current options from the
original crashed compile, however this is currently not possible.
Although a copy constructor exists, it is not a good idea to use it
here due to a number of problems:

1. The non-copy constructor takes care of initializing tracing, among
   other things, which may have not been active on the original
   compilation. There is currently no easy way for us to reinitialize
   such logic.

2. Options can be set at any point during the compilation, and
   decisions to set options can be based off of whether other options
   are set or not. This is very problematic. For example during the
   original compilation we may have set option A at some point X
   during the compilation thus changing the value of that option from
   that point onward. This means that if we were to copy construct
   the set of options from the original compile right here for the
   JitDump compilation then option A would yield a different value
   from the start of the compilation until point X.

This should eventually be improved if the options framework is ever
simplified to contain only option values, not things like
optimization plans, start PCs, compile thread IDs, etc.

Because of this limitation we do our best to only copy options which
are known to be timing sensitive and that could change between the
JitDump compilation and the original compilation. This is not a
silver bullet, and should be updated as we encounter more sources of
non-determinism for JitDump recompilation.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>