Fix StringPeephole optimization class redefinition check
The class in question may not have been loaded yet. For example in JDK11
the JCL removed most (all?) uses of the StringBuffer class, thus this
class is never loaded until `main` is reached. If the class is not
loaded it could not have been redefined, so we return false here.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>