Iterate thunk archetype with state for all passes of targeted inlining
Iterating method bytecode with state is able to get receiver info for
MethodHandle calls, thus able to find the target to inine. Currently,
only the first pass of targeted inlining does this. For long
MethodHandle chains that can't be inlined in one pass (due to the
inability of inliner to evaluate ILGen macros), more inlining passes
are required. The self-requesting opt is done in a recursive way, which
may result in stack overflow. Moreover, we should inline as many
MethodHandle as possible to save compilation cost.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>