Add support for jitdump agent sub-options
Sometimes, no matter how hard we try to heuristically determine what
tracing or options to enable for jitdump compiles, it may not be enough
diagnostics to determine the root cause of the problem. Manual
intervention is required in such cases. Currently this involves setting
-Xjit: options to add additional tracing and hoping to reproduce the
problem.

The current approach has several limitations:

- Tracing a method unconditionally (before a problem ever happens) can
make the problem completely stop happening due to the non-deterministic
nature of the JIT compiler

- Tracing with -Xjit: will always generate a trace file, even when no
problem happens

- Tracing with -Xjit: requires specifying the exact method to trace, or
a class of methods which can be problematic if the error at hand happens
in a variety of different methods, or even generated methods with unique
names

To overcome this limitation we can use the `opts=` sub-option of the
-Xdump command [1] which allows the user to specify any sub-options to
a particular command. We parse this command during the jitdump process
and we interpret it as a user specified method which we will trigger a
JitDump recompilation.

The user can specify a full method signature, for example:

```
-Xdump:jit:
    events=throw,
    filter=*NullPointerException#java/util/Calendar.setTime,
    range=1..1,
    opts=java/lang/String.indexOf(II)I
```

In the above example we tell the dump agent to trigger a JitDump when a
`NullPointerException` is thrown in the `java/util/Calendar.setTime`
method, and when that happens we want to recompile the method specified
in the `opts=` option, i.e. `java/lang/String.indexOf(II)I`.

The purpose of this sub-option is to debug non-deterministic style
problems described previously. Using this JitDump sub-option avoids the
non-determinism introduced by attempting to trace methods during their
first compilation.

[1] https://www.eclipse.org/openj9/docs/xdump/#optsltoptionsgt

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>