Acquire the write mutex earlier when updating runtime full flags
1. The existing code updates the runtime full flags and then acquires
the write mutex to change memory protection of partial filled pages. It
is possible that another thread could win the race of acquiring the
write mutex. It will start writing to the SCC without the memory
protection been changed, resulting in seg fault. Move the code of
acquiring the write mutex to the beginning of updateRuntimeFullFlags()
before the runtime full flags are updated.

2. Update Trc_SHR_CC_changePartialPageProtection_NotDone to include more
info.

Fixes #12015

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>