Allow AOT Warm compilations during startup under FSD
The initial change to enable AOT compilations under FSD did not allow
AOT warm compilations during startup as this improved startup. However,
this was only necessary if the SCC used is small; when using a bigger
SCC, the regression disappears and in fact, startup can even be
improved.

The reason why a larger SCC helps is because under FSD, the AOT bodies
are much bigger because of all the OSR code blocks. As a result, the SCC
could not hold as many AOT warm bodies as it can without FSD.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>