Make diagnostic thread compilations uninterruptible
It is possible that due to an external effect, for example running
out of JIT code cache, GC wanting to unload classes, etc. the
diagnostic thread compilation may be interrupted. We want to ensure
this does not happen by making compilations on the diagnostic thread
uninterruptible so that we can always generate a JitDump.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>