Clean up tree lowering for ArrayStoreCHK
Corrected lowerArrayStoreCHK comments to indicate that only awrtbari
is expected to be a child of an ArrayStoreCHK, and added a fatal
assertion to confirm that.

Moved generation of IL that examines the flags field of a class into
TR_J9VMBase to isolate optimizer from detailed knowledge of J9.

Removed a redundant call to copyByteCodeInfo.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>