Improve testJITServer test
In each test case of `testJITServer`, client and server
are terminated using `p.destroy()`. However, this call does not
guarantee that the process will exit. This commit increases
time a test will wait for the process to exit, and if it's still
running will terminate it with `p.destroyForcibly()` and fail the
test case.
This commit also enables to run all test cases even if the last
test case failed. Previously, if a test case fails before server is
terminated, then all subsequent cases fail as well, because they
cannot start their own server. Now, we check whether a previous
instance of a process is running, and if it is, we kill it first

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>