Remove code handling recusive crashes in JitDump
This should not be possible. The thread which crashed is guarded against
signals which are caught when the compilation is triggered. We cannot
encounter a recursive JitDump scenario because the dump mutex is held by
the crashed thread which has queued a compilation for the the JitDump
process to be handled by the diagnostic thread. Crashes on the
diagnostic thread are handled by simply returning. If we were to try to
do a recursive JitDump it would result in a deadlock at the JVM dump
handling level. As such the code here is dead and we can safely remove
it.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>