Associate class loaders with the name of first loaded class
In order to support JITServer AOT cache, we add a third index to the
class loader table - name of the first loaded class. Each class loader
is associated with the name of the first class it loaded. This mapping
will be used to find the right class loader to lookup classes in when
deserializing cached AOT methods received from the JITServer.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>