Stash the right number of arguments for invokehandle/invokedynamic
If the transition target is invokehandle/invokedynamic, the arguments to
be stashed are those already pushed onto the stack before generating the
bytecode, regardless of the number of arguments needed by the generated
call.

The generated call requires one (in resolved case) or two (in unresolved
case) implicit arguments, so we need to substract them to get the actual
number of arguments needed.

The reason why we don't need to stash the implicit arguments is because
when we transition to the VM, VM will push the implicit argument onto
the stack which is part of the behavior of the
invokehandle/invokedynamic bytecode.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>