Fix the API for destroying dynamically created monitors
This commit provides the correct implementation of the static method
TR::Monitor::destroy(TR::Monitor *monitor), which is the correct API
to destroy a monitor created with TR::Monitor::create().

The instance method TR::Monitor::destroy() should only be used for
monitors initialized with TR::Monitor::init() in cases when they are not
added to the MonitorTable and their storage is managed by the caller.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>