Add debug counters to track uses of value type helpers
This change adds new debug counters to aid in identifying areas in
generated code where many calls to value type helpers occur.  This
will help guide performance improvements with value types support
enabled.

The general form of the names of the counters is

vt-helper/generated/{acmp|aaload|aastore}/(<sig>)/bc=<idx>

where <sig> is the method signature and <idx> is the bytecode
index.

As optimizations are applied that eliminate calls to the helpers,
debug counters with the same general form will be used, except
that "generated" will be replaced with some indicator of the
optimization that was applied.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>