Handle aaload and aastore value type helpers in VP
Added changes to Value Propagation to allow helper calls for array
operations to be replaced with inline IL to perform the array access if
the array is known not to have a component type that is a value type.

Also, added debug counters for aaload, aastore and acmp Value
Propagation optimizations.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>