Suspend diagnostic thread after JitDump
If the user requested a JitDump via an event then the JVM continues
execution normally, however the diagnostic thread remains active. This
is not desirable because the diagnostic thread is looping waiting for
work and can start picking up non-JitDump compilation requests which
will cause an assert.

To prevent this and adhere to the contract outlined in `processEntries`
we suspend the diagnostic thread once the JitDump process is complete.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>