multianewArrayEvaluator segmentation fault fix
We were seeing a segmentation fault trying to init 2nd dim class. This
was because the memory reference to the class slot was invalid.

The memory reference is calculated using `targetReg` and `temp1Reg`.
`targetReg` seemed to be fine. However, `temp1Reg` carries over some
old garbage. When we write to `temp1Reg` using `SLLK` we only copy
lower 32 bits and leave the upper 32 bits untouched. This garbage in
upper 32 bits of `temp1Reg` is what was messing up our calculate of the
2nd dim class slot.

Changing `SLLk` to `SLLG` and `AFI` to `AGFI` fixes the failure.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>