Compile JitDump from crashed PC for application thread crashes
The stack walker may not be able to walk the stack if the crash did
not happen on a transition frame. In such cases the stack walker
will resume walking the stack on the last known valid point, which
will be a transition frame further up the stack (ex. INT -> JIT
transition frame). This will result in the stack walker potentially
skipping some JIT methods on the backtrace. This is not desirable.
Chances are high that the crash happen because of a miscompilation
in the first JIT compiled method on the stack, so it is imperative
that we recompile the method we actually crashed in.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>