Refactor IncrementalGenerationalGC/CopyForward code
Refactor VLHGC collector code in preparation for concurrent
Copy Forward collector. This is the first step of a series of
steps in order to introduce the low-pause collector that’s
based on the balanced GC policy.

Here we divide pre/post GC processing related to Copy
Forward. We also introduce concurrentCopyForwardCollectionSet
as well as isConcurrentCycleInProgress which will be used
by concurrent copy forward.

Enable concurrent Copy Forward in Linux x86 platforms.

This is phase 1 of #12089

Same PR as #12187 with Windows 32bit fix

Signed-off-by: Igor Braga <higorb1@gmail.com>