Add splitForFastpath helper and simplify code
This helper function splits a basic block at given point, assuming a
fastpath was just inserted. It also adds a CFG edge for the fastpath.
Conceptually, it takes care of the block/CFG changes involved in adding
a fastpath, significantly reducing some of the boiler plate code
that's needed to do so.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>