Define AOT serialization records and serialized AOT format
This commit adds the definitions of serialization records used in
JITServer AOT cache. Serialization records will be created for AOT
methods compiled and cached at the JITServer. These records store
information needed to update an AOT method body such that it can be
loaded into another client JVM running on a different host with a
different SCC than the JVM that the method was originally compiled for.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>