AArch64: Properly find compressed refs sequence in implicit NULLCHK
`evaluateNULLCHKWithPossibleResolve` does pattern matching to
find load operation in compressed references sequences, but it searches for
`TR::i2l` instead of correct opcode `TR::iu2l`.
Also, if it is a load operation and the reference count of the load node
under the compressed refs sequence is 1, explicit NULLCHK is used
even if the reference count of `TR::l2a` node is more than 2.

This commit changes `evaluateNULLCHKWithPossibleResolve` to
- Use opcode value of the parent of the reference node for judging
   whether it is a load operation instead of searching `TR::iu2l`.
- Use implicit NULLCHK if `TR::l2a` node's reference count > 2.
- Use recursivelyDecReferenceCount for firstChild if it is not evaluated.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>