Fix high number of live register in arrayCopyEvaluator
In arrayCopyEvaluator, the P10 version of helper uses VSX registers
that are not marked dead after usage. Which leads to incrementally
higher number of VSX registers being used until a memory corruption
happens. This commit fixes this bug by marking the VSX register as
dead. Also, to avoid high number of live FPR register, this commit
mark the live register as dead in "genWrtbarForArrayCopy" method.

This issue appears when gc uses concurrentScavenger.

Signed-off-by: Mohammad Nazmul Alam <mohammad.nazmul.alam@ibm.com>