Fix JitDump timing hole when resuming diagnostic thread
There is a small timing hole which I haven't observed yet but it is
possible that in the time between resuming the diagnostic thread and
purging the method queue we might pick up an OrdinaryMethod compile
on the diagnostic thread.

The fix here is to simply purge the queue and then enable the
diagnostic thread.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>