Fix a memory leak when caching remote ROM classes
JITServerHelpers::cacheRemoteROMClass attempts to store a newly allocated
remote ROM class into the cache. However, if after the class has been
requested but before it was cached, it was cached by another thread,
the new class will not be cached, resulting in a memory leak.
This commit fixes the problem by checking whether
`JITServerHelpers::cacheRemoteROMClass`
has actually cached the class and freeing the memory if it hasn't.

Closes: #12180

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>