Enable JitDump support on JITServer
This commit allows server to produce a JitDump log when a crash occurs
during a compilation.

Once the crash happens, the JitDump handler runs normally, producing
an IL log of the crashed thread until it reaches the point when
the problematic method needs to be recompiled. There, it sends
the client a `compilationThreadCrashed` message, asking it to schedule
a JitDump recompilation and blocks until the recompilation is done.

When the client receives the error message, it temporarily stops all
further compilations, purges the compilation queue and sends a request
to the server to recompile the method as a JitDump compilation.
The server will compile the method on the diagnostic thread, notify
the waiting crashed thread, allowing the JitDump handler to finish and
terminate the server.

Once the diagnostic compilation is complete, the client will resume
compilations locally until another server comes up.

Closes: #12394

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>