Unify handling of compiler options in JIT build with the rest of VM
This commit changes the order in which OMR_PLATFORM_COMPILE_OPTIONS and
OMR_PLATFORM_C[XX]_COMPILE_OPTIONS are added to the list of compiler
options in the JIT's CMakeLists.txt to match the rest of the VM build.

This allows the user to pass extra compiler options via
OMR_PLATFORM_C_COMPILE_OPTIONS and OMR_PLATFORM_CXX_COMPILE_OPTIONS
to override the defaults specified by OMR_PLATFORM_COMPILE_OPTIONS,
e.g. in order to produce a debug build.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>