Fix Z ArrayStoreCHKEvaluator to use a 64bit compare for compressedrefs
The Z ArrayStoreCHKEvaluator was using a 32bit compare of a decompressed
object reference with NULL when fast-pathing the write barrier. This
meant that when an object reference had all 0 bits in the lower 32bits
it would skip the write barrier and fail to mark the tenured object as
having a nursery reference. The end result is a possible early free of
nursery objects with the address 0x100000000 or 0x200000000 for example.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>