Fix invalid memory access when getting a ROM class
PR #12180 introduced a bug: when we try to cache a newly
received ROM class but find that it's already been cached,
we deallocate it but still return the pointer to the ROM class
that now contains invalid memory.
The fix is to update the class pointer with the cached value
and return that.

Closes: #12468, #12442

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>