Remove the appendix object when JIT calls linkToStaticSpecial
JIT will call linkToStatic for unresolved invokedynamic and invokehandle
sequences. For invokedynamic and invokehandle, the appendix object should not
be appended to the stack if it is null. For the unresolved cases, the JIT does
not know the value of the appendix object prior to calling linkToStatic. For
the JIT paths, the appendix object should be removed from the stack if the
_sendMethod does not expect it. The MemberName object is located at the top
of the stack. The appendix object is located before the MemberName object.

Shifting by 2 during the memmove will remove the appendix object.

Depends: #12549

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>