Fix fatal assertion when converting pointers to SCC offsets
Using TR_J9SharedCache::offsetInSharedCacheFromPointer() in AOT
deserializer to convert pointers such as class chains returned by
rememberClass() to SCC offsets can trigger a fatal assertion due
to the following race condition. When Thread 1 stores a class chain,
Thread 2 can look it up before Thread 1 updates the SCC area bounds
in the SCC header, and converting the pointer to offset in Thread 2
fails with a fatal assertion.

This commit is a workaround to avoid the fatal assertion by calling
TR_J9SharedCache::isPointerInSharedCache() which checks if the pointer
is in the SCC according to the current area bounds without a fatal
assertion, and failing deserialization if this call does not succeed.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>