Stop using unsafe sprintf() in InterpreterEmulator
In particular, in the debug tracing. The much safer snprintf() is now
used instead, though with the usual caveat that only _snprintf() (with
the underscore) is available in older versions of MSVC.

The implementations of Operand::printToString() are moved into
InterpreterEmulator.cpp to avoid putting #define snprintf ... into a
header file.

A wrapper is added that checks the result and prints a warning to the
log when truncation is detected. It also ensures NUL-termination in case
we are using _snprintf().

Call sites are updated to call the wrapper instead of calling the method
directly, and to use a larger buffer size. The old sizes were either
barely large enough (50), or definitely too small (20). The new size is
defined as a constant OPERAND_STR_BUF_SIZE to keep all of the call sites
in sync.