Check for arch support when generating AGHIK instruction
The multiANewArrayEvaluator in the z/codegen generates an
AGHIK instruction. This instruction is not supported on Z10
and lower platforms. This commit adds a check in the evaluator
to make sure that the current architecture supports the instruction
before generating it.

Tests for the evaluator have also been improved to account for all
supported architectures.

Fixes: #12538

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>