Thread iterator inside thread list lock
In Concurrent Scavenger, when close to terminating concurrent phase and
signaling mutator threads to flush their copy caches, prior to thread
iteration thread lock is acquired. However, it's taken after the
iterator construction, which itself advances the iteration to main VM
thread. The main thread may race and terminate before we acquire the
lock and start iteration.

The solution is to construct the iterator after acquiring the lock.


Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>